refactor: HotelEntity의 좌표 BigDecimal에서 Double로 변경

+++ src/main/java/staysplit/hotel_reservation/hotel/dto/request/UpdateHotelRequest.java
@@ -5,8 +5,8 @@ import java.math.BigDecimal;
 public record UpdateHotelRequest(
         String name,
         String address,
-        BigDecimal longitude,
-        BigDecimal latitude,
+        Double longitude,
+        Double latitude,
         String description,
         Integer starLevel
 ) {}
